+++ src/tab_cli/config.py
@@ -8,6 +8,7 @@ class Config:
     """Global configuration settings."""
 
     az_url_authority_is_account: bool = False
+    sampling_size_for_schema_inference: int = 32
 
 
 # Global config instance